Move SymPyDeprecationWarning into sympy.core.exceptions.
