+++ sympy/core/expr.py
@@ -4,7 +4,8 @@ from singleton import S
 from evalf import EvalfMixin, pure_complex
 from decorators import _sympifyit, call_highest_priority
 from cache import cacheit
-from compatibility import reduce, SymPyDeprecationWarning
+from compatibility import reduce
+from exceptions import SymPyDeprecationWarning
 from sympy.mpmath.libmp import mpf_log, prec_to_dps
 
 from collections import defaultdict
