+++ sympy/utilities/iterables.py
@@ -3,7 +3,8 @@ import random
 
 from sympy.core import Basic, C
 from sympy.core.compatibility import is_sequence, iterable #logically, these belong here
-from sympy.core.compatibility import product as cartes, combinations, combinations_with_replacement, SymPyDeprecationWarning
+from sympy.core.compatibility import product as cartes, combinations, combinations_with_replacement
+from sympy.core.exceptions import SymPyDeprecationWarning
 
 def flatten(iterable, levels=None, cls=None):
     """
